+++ NEWS.md
@@ -1,4 +1,4 @@
-# Version (development version)
+# Version 0.19.0 (2022-10-15)
 
 ## New Features